Decoupled Major Pingdom components into an adapter.
The project is moving to a modular design where one defines an adapter
and the adapter communicates to the dashboard what checks are up and are
down. This allows us to move away from one dashboard per monitoring
setup to a unified dashboard for multiple monitoring solutions.

This commit moves the key features of the Pingdom monitor out of the
main project itself to make the project open to various monitoring
platforms by creating new adapters for each monitor.