Added manual adapter, Migrated to configparser for monitor configs
The manual monitoring adapter exposes an api where you can set the
status of a predefined check using a web call. This monitoring adapter
is currently for testing purposes but could be expanded upon to create a
very simple monitor. A sample sams.conf file is also included that will
show users how monitors can be setup.

The pingdom adapter was modified to match the changes to the system so
far. I will be working on migrating the pingdom monitor to the new
adapter system to make the reporting system as a whole unified.